+++ CHANGELOG.rst
@@ -4,6 +4,19 @@
 
  Added
  -----
+ - Added mypy for linting
+ - Started adding config and parser
+ - Added _version and __main__ files
+ Changed
+ -------
+ - Updated files with license
+ - Updated versioning
+
+2023-01-06
+==========
+
+ Added
+ -----
  - Flake8 max-line-length=88
  - Starting to add Sphinx stuff
  Changed